% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regreg}
\alias{regreg}
\title{GOV.UK register register.}
\format{
An object of class \code{data.frame} with 51 rows and 10 columns.
}
\usage{
regreg
}
\description{
A dataset containing the characteristics of all
the published GOV.UK registers.
}
\details{
format: A data frame
\describe{
\item{register}{name of register}
\item{text}{description of register contents}
\item{registry}{responsible body of register}
\item{phase}{The stafe of development}
\item{copyright}{The copyright terms}
...
}
source: \url{https://webarchive.nationalarchives.gov.uk/ukgwa/20210104111014/https://www.registers.service.gov.uk/registers/register/download}
}
\keyword{datasets}
